add websocket functions to send data to remote server
Former-commit-id: 71c860cd61ba5550dd833fb4ba58abb164e1aeb9 [formerly a6e7938600ef5519915375fc7c7eef43d6086e7d] [formerly 63b4569604c588d8b9792ee751dd2e0614d4a470 [formerly 655168d4b94eb130d1f3de8cc50d9d5d55b508fe]]
Former-commit-id: ab759d9bfd450de0e2e928ddc4bfb794aab98431
Former-commit-id: df45b408fc37a0ae71907d7166abbc210404ba31

+++ iOS/Shybot/Shybot/CustomFaceRecognizer.mm
@@ -73,7 +73,7 @@
     
     sqlite3_finalize(statement);
     
-    return sqlite3_last_insert_rowid(_db);
+    return (int)sqlite3_last_insert_rowid(_db);
 }
 
 - (NSMutableArray *)getAllPeople
